Move createSupportTicket() and createTicketAnswerRecieved()
